+++ familyhelper-clannad-node/src/main/java/com/dwarfeng/familyhelper/clannad/node/configuration/CacheConfiguration.java
@@ -48,6 +48,8 @@ public class CacheConfiguration {
     private String notifyTopicPrefix;
     @Value("${cache.prefix.entity.notify_preference}")
     private String notifyPreferencePrefix;
+    @Value("${cache.prefix.entity.notify_meta}")
+    private String notifyMetaPrefix;
 
     public CacheConfiguration(RedisTemplate<String, ?> template, Mapper mapper) {
         this.template = template;
@@ -182,8 +184,19 @@ public class CacheConfiguration {
     notifyPreferenceRedisBatchBaseCache() {
         return new RedisBatchBaseCache<>(
                 (RedisTemplate<String, FastJsonNotifyPreference>) template,
-                new NotifyPreferenceStringKeyFormatter(notifyPreferencePrefix),
+                new NotifyNodeStringKeyFormatter(notifyPreferencePrefix),
                 new DozerBeanTransformer<>(NotifyPreference.class, FastJsonNotifyPreference.class, mapper)
+        );
+    }
+
+    @Bean
+    @SuppressWarnings("unchecked")
+    public RedisBatchBaseCache<NotifyNodeKey, NotifyMeta, FastJsonNotifyMeta>
+    notifyMetaRedisBatchBaseCache() {
+        return new RedisBatchBaseCache<>(
+                (RedisTemplate<String, FastJsonNotifyMeta>) template,
+                new NotifyNodeStringKeyFormatter(notifyMetaPrefix),
+                new DozerBeanTransformer<>(NotifyMeta.class, FastJsonNotifyMeta.class, mapper)
         );
     }
 }