+++ CHANGELOG.md
@@ -31,6 +31,7 @@
   - com.dwarfeng.familyhelper.clannad.stack.bean.entity.NotifySetting。
   - com.dwarfeng.familyhelper.clannad.stack.bean.entity.NotifyTopic。
   - com.dwarfeng.familyhelper.clannad.stack.bean.entity.NotifyPreference。
+  - com.dwarfeng.familyhelper.clannad.stack.bean.entity.NotifyMeta。
 
 - 优化操作服务验证环节的代码结构。